refactor: TxConfig로 TransactionalManager 설정 분리 반영
